+++ examples/src/main/java/CreateAccountWithAliasExample.java
@@ -78,10 +78,6 @@ public class CreateAccountWithAliasExample {
          * Step 6
          * Get the `AccountInfo` and show that the account has contractAccountId
          */
-        // `AccountInfoQuery` also queries the mirror node.
-        // Wait until the mirror node updates with the new data.
-        Thread.sleep(5000);
-
         AccountInfo accountInfo = new AccountInfoQuery()
             .setAccountId(newAccountId)
             .execute(client);